backend/payment_service/pyproject.toml
[project]
name = "payment-service"
version = "0.1.0"
description = "Stripe payment processing microservice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache>=0.12.3",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.14",
    "faststream[cli,rabbit]>=0.6.5",
    "gunicorn>=23.0.0",
    "orjson>=3.11.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "stripe>=15.0.1",
    "uvicorn>=0.35.0",
    "watchfiles>=1.1.0",
]


[tool.uv.sources]
shared = { path = "../shared", editable = true }

[tool.basedpyright]
extraPaths = [".."]
